fix issues in proc flag changes

+++ engine/item/special_effect.hpp
@@ -45,7 +45,7 @@ struct special_effect_t
   special_effect_source_e source;
   std::string name_str, trigger_str, encoding_str;
   uint64_t proc_flags_; /* Proc-by */
-  unsigned proc_flags2_; /* Proc-on (hit/damage/...) */
+  uint64_t proc_flags2_; /* Proc-on (hit/damage/...) */
   stat_e stat;
   school_e school;
   int max_stacks;
@@ -154,7 +154,7 @@ public:
 
   /* Accessors for driver specific features of the proc; some are also used for on-use effects */
   uint64_t proc_flags() const;
-  unsigned proc_flags2() const;
+  uint64_t proc_flags2() const;
   double ppm() const;
   double rppm() const;
   unsigned rppm_scale() const;
